Add sound resources
These are WAV versions of AIFF files used in quotidian.  WAV is a
generally compatible format, usable in Firefox, Chrome, and Safari.
This MicroSoft-created file format works in Edge, bit not Internet
Explorer.

The silence.aiff file was there to work around a MacOS NSSound bug and
was not copied in the hope it will not be needed in Spar.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>