Add remainderText to create a new function to derive remainder labels
There are several places where it would be nice to produce Strings of
the same format.  As this format is set only once at the beginning,
a function with all but one variable fixed after that point should be
simpler to manage.  And I wanted to try mixing a partial application,
an arrow function, and a closure together.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>