Improve timer height calculation
Tries to fix the readability of this calculation by splitting most of it
into a diameter measurement and a final radius determination.

This also limits the maximum size to 95% of the available space so that
scroll bars are not needed.  The "overflow: hidden" setting is retained to
doubly ensure no scroll bars are shown.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>