Use ECMAScript 6 for coverage tests
According to the documentation for Blanket.js, it should be possible to
use 'data-cover-ecmaVersion="6"' on the script tag for spar.js.  But, I've
not been able to get this to work and I cannot see where this
functionality is included in the blanket.js file.  So, as a work-around,
change the default 'ecmaVersion' in the script.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>