Add tests for getParams and fix found bugs
Test coverage measurement does not like for...of and for...in is the
wrong choice, so use a standard for loop and the URLSearchParams
forEach method.  for...in includes all enumerable properties, even those
we don't care about in this milieu.

With working tests, two bugs showed themselves.  First, when the default
value for options[name] was false, the URL search parameter was treated
as an unknown option.  Second, unrecognized options were logged, but
still added to the params object returned to the caller.  Now, we delete
undefined parameters before returning the constructed object.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>